Move types to avoid circular dep issue
